requirements.txt
numpy
pandas
ripser
persim
networkx
scikit-learn
statsmodels
yfinance
matplotlib
pytest
requests
python-dotenv
websockets
